+++ src/Classifiers/SoftmaxClassifier.php
@@ -73,10 +73,9 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
     protected float $l2Penalty;
 
     /**
-     * The maximum number of training epochs. i.e. the number of times to iterate
-     * over the entire training set before terminating.
+     * The maximum number of training epochs. i.e. the number of times to iterate before terminating.
      *
-     * @var int
+     * @var int<0,max>
      */
     protected int $epochs;
 
@@ -90,7 +89,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
     /**
      * The number of epochs without improvement in the training loss to wait before considering an early stop.
      *
-     * @var int
+     * @var positive-int
      */
     protected int $window;
 
@@ -151,7 +150,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
                 . " greater than 0, $l2Penalty given.");
         }
 
-        if ($epochs < 1) {
+        if ($epochs < 0) {
             throw new InvalidArgumentException('Number of epochs'
                 . " must be greater than 0, $epochs given.");
         }